Update with targets to cleanup files
This is an update to the `2-Makefile` makefile. It adds the following
targets, `clean` for temporary file cleanups, `oclean` for object file
cleanup, and `fclean` for a full cleanup of both temporary and object
files. Also it adds the `re` target which forces a recompilation and
building of the executable file.